third_party/blink/web_tests/traversal/node-iterator-001-expected.txt

Expect ref node root; pointer after: A B C D E F G H I

Expect ref node 'A'; pointer after: [A] * B C D E F G H I

Expect ref node 'A'; pointer before: * [A] B C D E F G H I

Expect ref node 'A'; pointer before: * [A] B C D E F G H I
